// src/interfaces/collection.interface.ts
import { IUser } from './user.interface';

export interface Collection {
    _id: string;
    name: string;
    role: number;
    chain: string;
    contract_address: string;
    contract_name: string;
    logo_url: string;
    banner_url: string;
    items_total: number;
    owners_total: number;
    sales_1d: number;
    sales_7d: number;
    sales_30d: number;
    sales_total: number;
    volume_1d: number;
    volume_7d: number;
    volume_30d: number;
    volume_total: number;
    floor_price: number;
    followersCount:number;
    average_price_1d: number;
    average_price_7d: number;
    average_price_30d: number;
    average_price_total: number;
    average_price_change_1d: number;
    average_price_change_7d: number;
    average_price_change_30d: number;
    volume_change_1d: number;
    volume_change_7d: number;
    market_cap: number;
    symbol: string;
    description: string;
    website: string;
    email: string;
    twitter: string;
    discord: string;
    telegram: string;
    github: string;
    instagram: string;
    medium: string;
    featured_url: string;
    large_image_url: string;
    attributes: any[];
    erc_type: string;
    deploy_block_number: number;
    owner: string | IUser | any;
    verified: boolean;
    opensea_verified: boolean;
    royalty: number;
    amounts_total: number;
    collections_with_same_name: any[];
    price_symbol: string;
    followers: string[];
    favourites: string[];
    likes: string[];
    views: string[];
}